# backend/app/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: SNS Conference Tool API
  description: |
    REST API for the SNS Conference Tool. Pass 1 (current): implemented by MSW
    handlers in `web/lib/api/mocks/handlers.ts`. Pass 2: implemented by the
    Spring Boot backend in this module.

    This specification is the single source of truth for request/response
    shapes. When a real backend endpoint replaces a mock, the MSW handler
    is deleted and only this schema remains.
  version: 0.1.0
servers:
  - url: /api
    description: Same-origin (default for the WebView)
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Start registration (email only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202":
          description: Accepted; verification email dispatched
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted: { type: boolean }
  /auth/verify:
    post:
      tags: [Auth]
      summary: Verify email with TAN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, tan]
              properties:
                email: { type: string, format: email }
                tan: { type: string, minLength: 6, maxLength: 6 }
      responses:
        "200":
          description: TAN accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified: { type: boolean }
                  verificationToken: { type: string }
  /auth/complete:
    post:
      tags: [Auth]
      summary: Complete registration after verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verificationToken, firstName, lastName, password]
              properties:
                verificationToken: { type: string }
                firstName: { type: string }
                lastName: { type: string }
                academicTitle: { type: [string, "null"] }
                institution: { type: [string, "null"] }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Registration complete, tokens issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
  /auth/login:
    post:
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
  /auth/refresh:
    post:
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New access+refresh pair (refresh token is rotated)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
  /auth/logout:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Logged out; refresh token revoked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, userId]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        userId: { type: string, format: uuid }
...